backend/requirements.txt
fastapi
uvicorn
pypdf
groq
python-multipart
python-dotenv